requirements.txt
docker==6.1.3
PyYAML==6.0.1
click==8.1.7
cvss==2.6
#guesslang==2.2.1
#python-magic==0.4.27 